requirements.txt
fastapi==0.68.1
flake8==3.9.2
mysql-connector-python==8.0.26
pandas==1.3.3
pipdeptree==2.2.0
python-dotenv==0.19.0
uvicorn==0.15.0
